ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pillow==10.1.0
openai==1.3.0
anthropic==0.7.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
torch>=2.2.0
transformers>=4.49.0
accelerate>=0.25.0
peft>=0.7.0
qwen-vl-utils
bitsandbytes>=0.41.3
